requirements.txt
pymongo[srv]
flask
python-dotenv